Move the document session and apply into useAgentBuilderForm
The hook is the only thing that applies drafts to the document, so
agentConfigBasicYaml.ts folds into it: session creation, the
diff-and-rewrite apply, and the wire serializers now live with the form
that uses them. The two test files and their fixtures merge into
useAgentBuilderForm.test.ts. agentConfigBasicExtract.ts stays separate
as the schema-validated read side (folding it in would break the
500-line lint limit).
